Refactor out item size functions and improve a little
